Fixed a mistake, added packaged main.java.Duke
